+++ gui/headers/Point.h
@@ -7,8 +7,8 @@ public:
     Point(double x, double y);
     double x();
     double y();
-    void setX(int x);
-    void setY(int y);
+    void setX(double x);
+    void setY(double y);
     void glDraw();
 
 protected:
